M    lynx/Obj.cpp M    lynx/GameObjPlayer.h M    lynx/ClientHUD.cpp M    lynx/ParticleSystemBlood.h M    lynx/ParticleSystemDust.h M    lynx/main.cpp M    lynx/WorldClient.h M    lynx/lynx.h

+++ lynx/lynx.h
@@ -3,11 +3,13 @@
 #include <assert.h>
 #include <string>       // Default String Type for Lynx
 #include <stdint.h>
+#include <cstdlib>
+#include <limits.h>
 
-typedef unsigned long   DWORD;
-typedef int             INT32;
-typedef short           INT16;
-typedef unsigned short  WORD;
+typedef uint32_t        DWORD;
+typedef int32_t         INT32;
+typedef int16_t         INT16;
+typedef uint16_t        WORD;
 typedef unsigned char   BYTE;
 
 #define SAFE_RELEASE(x) if(x) { delete x; x = NULL; }
